Rectangle and Square
You have a rectangle of length 
A
A units and breadth 
B
B units. You also have a square with side length of 
C
C units.

Check if both the rectangle and the square have the same area. Print 
Yes
Yes if they have the same area, and 
No
No otherwise.

Input Format
The first and only line of input contains 
3
3 integers - 
A
A, 
B
B and 
C
C.
Output Format
Output on a new line 
Yes
Yes if they have the same area, and 
No
No otherwise.

It is allowed to print each character in either case, i.e. 
YES
YES, 
yes
yes, 
yES
yES will all be accepted as a positive response.

Constraints
1
≤
A
,
B
,
C
≤
10
1≤A,B,C≤10
Sample 1:
Input
Output
2 8 4
Yes
Explanation:
Both the rectangle and the square have an area of 
16
16 units
2
2
 .

Sample 2:
Input
Output
3 5 4
No
Explanation:
The rectangle has an area of 
15
15 units
2
2
  but the square has 
16
16 units
2
2
 .