Equalize AB ( Difficulty:1069)
You are given two numbers 
A
A and 
B
B along with an integer 
X
X. In one operation you can do one of the following:

Set 
A
=
A
+
X
A=A+X and 
B
=
B
−
X
B=B−X
Set 
A
=
A
−
X
A=A−X and 
B
=
B
+
X
B=B+X
Determine if you can make 
A
A and 
B
B equal after applying the operation any number of times (possibly zero).

Input Format
The first line contains a single integer 
T
T — the number of test cases. Then the test cases follow.
The first and only line of each test case contains two space-separated integers 
A
,
B
A,B and 
X
X — the parameters mentioned in the statement.
Output Format
For each test case, output YES if you can make 
A
A and 
B
B equal after applying the operation any number of times (possibly zero). Otherwise, output NO.

You can output each letter in any case i.e. YES, yes, yEs are all considered the same.

Constraints
1
≤
T
≤
1000
1≤T≤1000
1
≤
A
,
B
,
X
≤
10
9
1≤A,B,X≤10 
9
 
Sample 1:
Input
Output
4
5 7 1
3 4 2
4 4 6
2 5 3
YES
NO
YES
NO
Explanation:
Test Case 1: The initial values of 
(
A
,
B
)
(A,B) is 
(
5
,
7
)
(5,7). We can perform the following operation: 
(
5
,
7
)
→
A
=
A
+
X
,
B
=
B
−
X
(
6
,
6
)
(5,7) 
A=A+X,B=B−X
​
 (6,6).

Test Case 2: It can be proven that we can not make 
A
A equal to 
B
B using the given operations.

Test Case 3: 
A
A and 
B
B are already equal.